Test and fix ipv4 address output.

+++ include/distrac/types.h
@@ -53,8 +53,10 @@ distrac_type_required_padding(enum distrac_type type);
 
 inline std::ostream&
 operator<<(std::ostream& o, const distrac_ipv4& ipv4) {
-  return o << ipv4.segments[0] << "." << ipv4.segments[0] << "."
-           << ipv4.segments[0] << "." << ipv4.segments[0];
+  return o << static_cast<int>(ipv4.segments[0]) << "."
+           << static_cast<int>(ipv4.segments[0]) << "."
+           << static_cast<int>(ipv4.segments[0]) << "."
+           << static_cast<int>(ipv4.segments[0]);
 }
 
 inline std::ostream&
